Introduced the render planner
The render planner is there to turn the game state into a render state.
This makes the actual rendering logic simpler, as all it needs to do is
to render the cells and segments that are specified in the render plan.

This basic version of the render planner simply ensures that all
segments of all cells are rendered, but segments are only rendered once.
It does not care what type of cell it is.